+++ libs/libc/sysdeps/unix/generic/ioctl.c
@@ -1,7 +1,7 @@
 #include <sys/ioctl.h>
 #include <sysdep.h>
 
-int ioctl(int fd, uint32_t cmd, uint32_t arg)
+int ioctl(int fd, uintptr_t cmd, uintptr_t arg)
 {
     int res = DO_SYSCALL_3(SYS_IOCTL, fd, cmd, arg);
     RETURN_WITH_ERRNO(res, res, -1);
